+++ drivers/usb/usbd/usbd.c
@@ -51,8 +51,8 @@ DriverEntry(PDRIVER_OBJECT DriverObject,
 /*
  * @implemented
  */
-DWORD NTAPI
-DllInitialize(DWORD Unknown)
+ULONG NTAPI
+DllInitialize(ULONG Unknown)
 {
     return 0;
 }
@@ -60,7 +60,7 @@ DllInitialize(DWORD Unknown)
 /*
  * @implemented
  */
-DWORD NTAPI
+ULONG NTAPI
 DllUnload(VOID)
 {
     return 0;
@@ -70,7 +70,7 @@ DllUnload(VOID)
  * @implemented
  */
 PVOID NTAPI
-USBD_Debug_GetHeap(DWORD Unknown1, POOL_TYPE PoolType, ULONG NumberOfBytes,
+USBD_Debug_GetHeap(ULONG Unknown1, POOL_TYPE PoolType, ULONG NumberOfBytes,
 	ULONG Tag)
 {
     return ExAllocatePoolWithTag(PoolType, NumberOfBytes, Tag);
@@ -80,7 +80,7 @@ USBD_Debug_GetHeap(DWORD Unknown1, POOL_TYPE PoolType, ULONG NumberOfBytes,
  * @implemented
  */
 VOID NTAPI
-USBD_Debug_RetHeap(PVOID Heap, DWORD Unknown2, DWORD Unknown3)
+USBD_Debug_RetHeap(PVOID Heap, ULONG Unknown2, ULONG Unknown3)
 {
     ExFreePool(Heap);
 }
@@ -98,7 +98,7 @@ USBD_Debug_LogEntry(PCHAR Name, ULONG_PTR Info1, ULONG_PTR Info2,
  * @implemented
  */
 PVOID NTAPI
-USBD_AllocateDeviceName(DWORD Unknown)
+USBD_AllocateDeviceName(ULONG Unknown)
 {
     return NULL;
 }
@@ -106,20 +106,20 @@ USBD_AllocateDeviceName(DWORD Unknown)
 /*
  * @implemented
  */
-DWORD NTAPI
+ULONG NTAPI
 USBD_CalculateUsbBandwidth(
     ULONG MaxPacketSize,
     UCHAR EndpointType,
     BOOLEAN LowSpeed
     )
 {
-    DWORD OverheadTable[] = {
+    ULONG OverheadTable[] = {
             0x00, /* UsbdPipeTypeControl */
             0x09, /* UsbdPipeTypeIsochronous */
             0x00, /* UsbdPipeTypeBulk */
             0x0d  /* UsbdPipeTypeInterrupt */
         };
-    DWORD Result;
+    ULONG Result;
 
     if (OverheadTable[EndpointType] != 0)
     {
@@ -134,8 +134,8 @@ USBD_CalculateUsbBandwidth(
 /*
  * @implemented
  */
-DWORD NTAPI
-USBD_Dispatch(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3, DWORD Unknown4)
+ULONG NTAPI
+USBD_Dispatch(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3, ULONG Unknown4)
 {
     return 1;
 }
@@ -167,8 +167,8 @@ USBD_WaitDeviceMutex(PVOID Unknown)
 /*
  * @implemented
  */
-DWORD NTAPI
-USBD_GetSuspendPowerState(DWORD Unknown1)
+ULONG NTAPI
+USBD_GetSuspendPowerState(ULONG Unknown1)
 {
     return 0;
 }
@@ -177,8 +177,8 @@ USBD_GetSuspendPowerState(DWORD Unknown1)
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_InitializeDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
-    DWORD Unknown4, DWORD Unknown5, DWORD Unknown6)
+USBD_InitializeDevice(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3,
+    ULONG Unknown4, ULONG Unknown5, ULONG Unknown6)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -187,9 +187,9 @@ USBD_InitializeDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_RegisterHostController(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
-    DWORD Unknown4, DWORD Unknown5, DWORD Unknown6, DWORD Unknown7,
-    DWORD Unknown8, DWORD Unknown9, DWORD Unknown10)
+USBD_RegisterHostController(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3,
+    ULONG Unknown4, ULONG Unknown5, ULONG Unknown6, ULONG Unknown7,
+    ULONG Unknown8, ULONG Unknown9, ULONG Unknown10)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -198,7 +198,7 @@ USBD_RegisterHostController(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_GetDeviceInformation(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
+USBD_GetDeviceInformation(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -207,8 +207,8 @@ USBD_GetDeviceInformation(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_CreateDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
-    DWORD Unknown4, DWORD Unknown5)
+USBD_CreateDevice(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3,
+    ULONG Unknown4, ULONG Unknown5)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -217,7 +217,7 @@ USBD_CreateDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3,
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_RemoveDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
+USBD_RemoveDevice(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -226,7 +226,7 @@ USBD_RemoveDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
  * @implemented
  */
 VOID NTAPI
-USBD_CompleteRequest(DWORD Unknown1, DWORD Unknown2)
+USBD_CompleteRequest(ULONG Unknown1, ULONG Unknown2)
 {
 }
 
@@ -245,7 +245,7 @@ USBD_RegisterHcFilter(
  * @implemented
  */
 VOID NTAPI
-USBD_SetSuspendPowerState(DWORD Unknown1, DWORD Unknown2)
+USBD_SetSuspendPowerState(ULONG Unknown1, ULONG Unknown2)
 {
 }
 
@@ -253,7 +253,7 @@ USBD_SetSuspendPowerState(DWORD Unknown1, DWORD Unknown2)
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_MakePdoName(DWORD Unknown1, DWORD Unknown2)
+USBD_MakePdoName(ULONG Unknown1, ULONG Unknown2)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -289,7 +289,7 @@ USBD_GetUSBDIVersion(
  * @implemented
  */
 NTSTATUS NTAPI
-USBD_RestoreDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
+USBD_RestoreDevice(ULONG Unknown1, ULONG Unknown2, ULONG Unknown3)
 {
     return STATUS_NOT_SUPPORTED;
 }
@@ -298,8 +298,8 @@ USBD_RestoreDevice(DWORD Unknown1, DWORD Unknown2, DWORD Unknown3)
  * @implemented
  */
 VOID NTAPI
-USBD_RegisterHcDeviceCapabilities(DWORD Unknown1, DWORD Unknown2,
-    DWORD Unknown3)
+USBD_RegisterHcDeviceCapabilities(ULONG Unknown1, ULONG Unknown2,
+    ULONG Unknown3)
 {
 }
 
@@ -315,8 +315,8 @@ USBD_CreateConfigurationRequestEx(
     )
 {
     PURB Urb;
-    DWORD UrbSize;
-    DWORD InterfaceCount;
+    ULONG UrbSize;
+    ULONG InterfaceCount;
 
     for (InterfaceCount = 0;
          InterfaceList[InterfaceCount].InterfaceDescriptor != NULL;
@@ -363,7 +363,7 @@ USBD_GetInterfaceLength(
 {
     ULONG_PTR Current;
     PUSB_INTERFACE_DESCRIPTOR CurrentDescriptor = InterfaceDescriptor;
-    DWORD Length = CurrentDescriptor->bLength;
+    ULONG Length = CurrentDescriptor->bLength;
 
     // USB_ENDPOINT_DESCRIPTOR_TYPE
     if (CurrentDescriptor->bDescriptorType == USB_INTERFACE_DESCRIPTOR_TYPE)
@@ -478,7 +478,7 @@ USBD_ParseConfigurationDescriptor(
 /*
  * @implemented
  */
-DWORD NTAPI
+ULONG NTAPI
 USBD_GetPdoRegistryParameter(
     PDEVICE_OBJECT PhysicalDeviceObject,
     PVOID Parameter,
