+++ lib/3rdparty/freetype/src/pcf/pcfutil.h
@@ -37,15 +37,15 @@ FT_BEGIN_HEADER
 
   FT_LOCAL( void )
   BitOrderInvert( unsigned char*  buf,
-                  int             nbytes );
+                  size_t          nbytes );
 
   FT_LOCAL( void )
   TwoByteSwap( unsigned char*  buf,
-               int             nbytes );
+               size_t          nbytes );
 
   FT_LOCAL( void )
   FourByteSwap( unsigned char*  buf,
-                int             nbytes );
+                size_t          nbytes );
 
 FT_END_HEADER
