+++ ntoskrnl/se/lsa.c
@@ -49,7 +49,7 @@ NTSTATUS
 NTAPI
 LsaFreeReturnBuffer(PVOID Buffer)
 {
-    ULONG Size = 0;
+    SIZE_T Size = 0;
     return ZwFreeVirtualMemory(NtCurrentProcess(),
                                &Buffer,
                                &Size,
@@ -110,8 +110,8 @@ NTSTATUS
 NTAPI
 SeMarkLogonSessionForTerminationNotification(IN PLUID LogonId)
 {
-	UNIMPLEMENTED;
-	return STATUS_NOT_IMPLEMENTED;
+    UNIMPLEMENTED;
+    return STATUS_NOT_IMPLEMENTED;
 }
 
 /*
@@ -121,8 +121,8 @@ NTSTATUS
 NTAPI
 SeRegisterLogonSessionTerminatedRoutine(IN PSE_LOGON_SESSION_TERMINATED_ROUTINE CallbackRoutine)
 {
-	UNIMPLEMENTED;
-	return STATUS_NOT_IMPLEMENTED;
+    UNIMPLEMENTED;
+    return STATUS_NOT_IMPLEMENTED;
 }
 
 /*
@@ -132,8 +132,8 @@ NTSTATUS
 NTAPI
 SeUnregisterLogonSessionTerminatedRoutine(IN PSE_LOGON_SESSION_TERMINATED_ROUTINE CallbackRoutine)
 {
-	UNIMPLEMENTED;
-	return STATUS_NOT_IMPLEMENTED;
+    UNIMPLEMENTED;
+    return STATUS_NOT_IMPLEMENTED;
 }
 
 /* EOF */
